+++ mvc/src/main/java/nextstep/mvc/handlerAdapter/ManualHandlerAdapter.java
@@ -1,8 +1,7 @@
-package nextstep.mvc.controller.tobe;
+package nextstep.mvc.handlerAdapter;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
-import nextstep.mvc.HandlerAdapter;
 import nextstep.mvc.controller.asis.Controller;
 import nextstep.mvc.view.JspView;
 import nextstep.mvc.view.ModelAndView;
